+++ src/handler/page.rs
@@ -190,12 +190,12 @@ impl PageInner {
     }
 
     /// Performs a mouse click event at the point's location
-    pub async fn click(&self, point: Point) -> Result<&Self> {
+    pub async fn click(&self, point: Point, options: chromiumoxide_types::ClickOptions) -> Result<&Self> {
         let cmd = DispatchMouseEventParams::builder()
             .x(point.x)
             .y(point.y)
             .button(MouseButton::Left)
-            .click_count(1);
+            .click_count(options.click_count);
 
         self.move_mouse(point)
             .await?