+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `add_init_script` to `Page` for scripts before navigation
+- Add `click_with` to `Page` and `Element` for custom click behavior
 
 ## [0.8.0] 2025-11-28